chore: move util functions to fixtures/
