+++ tests/integration/src/querier/04_traces.py
@@ -6,12 +6,12 @@ import requests
 
 from fixtures import types
 from fixtures.auth import USER_ADMIN_EMAIL, USER_ADMIN_PASSWORD
-from fixtures.traces import TraceIdGenerator, Traces, TracesKind, TracesStatusCode
-from src.querier.timeseries_utils import (
+from fixtures.querier import (
     assert_minutely_bucket_values,
     find_named_result,
     index_series_by_label,
 )
+from fixtures.traces import TraceIdGenerator, Traces, TracesKind, TracesStatusCode
 
 
 def test_traces_list(
